Draft: move functions to draftgeoutils.sort_edges
